feat: 파일명 수정 후 프리셋 모달 컴포넌트로 대체
